fix: complete packet loss type conversions (all Long to Int)
Converted packetsDropped and packetsFailed to Int to match recordPacketLoss signature.
Build now successful. WiFi telemetry fully working: battery, RSSI, RTT, and packet loss.

+++ app/src/main/java/com/metelci/ardunakon/ui/screens/ControlScreen.kt
@@ -88,8 +88,8 @@ fun ControlScreen(
                 bluetoothManager.telemetryHistoryManager.recordPacketLoss(
                     packetsSent = t.packetsSent.toInt(),
                     packetsReceived = (t.packetsSent - t.packetsDropped - t.packetsFailed).toInt(),
-                    packetsDropped = t.packetsDropped,
-                    packetsFailed = t.packetsFailed
+                    packetsDropped = t.packetsDropped.toInt(),
+                    packetsFailed = t.packetsFailed.toInt()
                 )
             } ?: android.util.Log.d("WiFiTelemetry", "wifiTelemetry is NULL")
         }
